requirements.txt
# --- LangChain Core Dependencies (Pinned for compatibility) ---
# Pinning langchain to a version before the major 0.1.0 update
# to ensure compatibility with modules like GenerativeAgent that use old import paths.
langchain==0.0.354
langchain-experimental==0.0.47

# --- Other Project Dependencies ---
anthropic
beautifulsoup4
chromadb
fastapi
python-multipart
# langchain and langchain_experimental are specified above
modelscope
openai
tiktoken
torch
tqdm
transformers
uvicorn
pillow
faiss-cpu
huggingface_hub[cli]
google-generativeai
pydantic
python-socketio[asyncio]
httpx[socks]  # 支持 SOCKS 代理（如果使用代理环境）